Add minimum working fips setup
a fips static lib was added to be able to have the slmem files as a single
entry in the generated project file, to make that work, a stub file was
created that does nothing but include the slmem.h header.

added tests group and renamed main.cpp to test0.cpp, I'll revisit it in a
further change to make the test within more uniform.

fixed a warning in alloc_debug when compiling for x64